@@ -32,10 +32,11 @@ from cvxpy.atoms.affine.partial_transpose import partial_transpose
 from cvxpy.atoms.affine.promote import promote
 from cvxpy.atoms.affine.real import real
 from cvxpy.atoms.affine.reshape import deep_flatten, reshape
+from cvxpy.atoms.affine.squeeze import squeeze
 from cvxpy.atoms.affine.concatenate import concatenate
 from cvxpy.atoms.affine.sum import sum
-from cvxpy.atoms.affine.trace import trace
-from cvxpy.atoms.affine.transpose import transpose
+from cvxpy.atoms.affine.trace import trace, Trace
+from cvxpy.atoms.affine.transpose import (transpose, permute_dims, swapaxes, moveaxis)
 from cvxpy.atoms.affine.upper_tri import upper_tri, vec_to_upper_tri
 from cvxpy.atoms.affine.vec import vec
 from cvxpy.atoms.affine.vstack import vstack
