@@ -113,6 +113,7 @@ class Leaf(expression.Expression):
                 raise ValueError("Invalid dimensions %s." % (shape,))
         shape = tuple(shape)
         self._shape = shape
+        super(Leaf, self).__init__()
 
         if (PSD or NSD or symmetric or diag or hermitian) and (len(shape) != 2
                                                                or shape[0] != shape[1]):
@@ -145,6 +146,7 @@ class Leaf(expression.Expression):
             self.integer_idx = integer
         if sparsity:
             self.sparse_idx = self._validate_indices(sparsity)
+            self._sparse_high_fill_in = (len(self.sparse_idx[0]) / np.prod(self.shape) <= 0.25)
         else:
             self.sparse_idx = None
         # count number of attributes
@@ -401,13 +403,15 @@ class Leaf(expression.Expression):
         elif self.attributes['complex']:
             return val.astype(complex)
         elif self.attributes['boolean']:
-            # TODO(akshayka): respect the boolean indices.
-            return np.round(np.clip(val, 0., 1.))
+            if hasattr(self, "boolean_idx"):
+                new_val = np.atleast_1d(val.astype(np.float64, copy=True))
+                new_val[self.boolean_idx] = np.round(np.clip(new_val[self.boolean_idx], 0., 1.))
+                return new_val.reshape(val.shape) if val.ndim == 0 else new_val
         elif self.attributes['integer']:
-            # TODO(akshayka): respect the integer indices.
-            # also, a variable may be integer in some indices and
-            # boolean in others.
-            return np.round(val)
+            if hasattr(self, "integer_idx"):
+                new_val = np.atleast_1d(val.astype(np.float64, copy=True))
+                new_val[self.integer_idx] = np.round(new_val[self.integer_idx])
+                return new_val.reshape(val.shape) if val.ndim == 0 else new_val
         elif self.attributes['diag']:
             if intf.is_sparse(val):
                 val = val.diagonal()
@@ -473,7 +477,7 @@ class Leaf(expression.Expression):
 
     @value.setter
     def value(self, val) -> None:
-        if self.sparse_idx is not None:
+        if self.sparse_idx is not None and self._sparse_high_fill_in:
             warnings.warn('Writing to a sparse CVXPY expression via `.value` is discouraged.'
                           ' Use `.value_sparse` instead', RuntimeWarning, 1)
         self.save_value(self._validate_value(val))
@@ -547,7 +551,7 @@ class Leaf(expression.Expression):
             # ^ might be a numpy array, scipy matrix, or sparse scipy matrix.
             if intf.is_sparse(delta):
                 # ^ based on current implementation of project(...),
-                #   is is not possible for this Leaf to be PSD/NSD *and*
+                #   it is not possible for this Leaf to be PSD/NSD *and*
                 #   a sparse matrix.
                 close_enough = np.allclose(delta.data, 0,
                                            atol=SPARSE_PROJECTION_TOL)
