@@ -22,6 +22,7 @@ import cvxpy.interface as intf
 import cvxpy.settings as s
 from cvxpy.error import SolverError
 from cvxpy.reductions.solution import Solution, failure_solution
+from cvxpy.reductions.solvers import utilities
 from cvxpy.reductions.solvers.conic_solvers.conic_solver import (
     ConicSolver,
     dims_to_solver_dict,
@@ -148,14 +149,26 @@ class HIGHS(ConicSolver):
         if status in s.SOLUTION_PRESENT:
             opt_val = results["info"].objective_function_value + inverse_data[s.OFFSET]
             primal_vars = {
+                # inverse_data[HIGHS.VAR_ID]: ...
+                # I don't understand how the line below works, the other conif solvers have
+                # something similar to the commented line above for the "key".
                 HIGHS.VAR_ID: intf.DEFAULT_INTF.const_to_matrix(
                     np.array(results["solution"].col_value)
                 )
             }
             # add duals if not a MIP.
             dual_vars = None
-            if not inverse_data["is_mip"]:
-                dual_vars = {HIGHS.DUAL_VAR_ID: -np.array(results["solution"].row_dual)}
+            if not inverse_data['is_mip']:
+                # The dual values are retrieved in the order that the
+                # constraints were added in solve_via_data() below. We
+                # must be careful to map them to inverse_data[EQ_CONSTR]
+                # followed by inverse_data[NEQ_CONSTR] accordingly.
+                y = -np.array(results["solution"].row_dual)
+                dual_vars = utilities.get_dual_values(
+                    y,
+                    utilities.extract_dual_value,
+                    inverse_data[HIGHS.EQ_CONSTR] + inverse_data[HIGHS.NEQ_CONSTR])
+
             attr[s.NUM_ITERS] = (
                 results["info"].ipm_iteration_count
                 + results["info"].crossover_iteration_count
@@ -257,14 +270,22 @@ class HIGHS(ConicSolver):
         lp.col_lower_ = col_lower
         lp.col_upper_ = col_upper
 
+        solver = hp.Highs()
+
         # setup options
         unpack_highs_options_inplace(solver_opts)
-        options = hp.HighsOptions()
-        options.log_to_console = verbose
-        for key, value in solver_opts.items():
-            setattr(options, key, value)
+        write_model_file = solver_opts.pop("write_model_file", None)
+        solver.setOptionValue("log_to_console", verbose)
+        for name, value in solver_opts.items():
+            # note that calling setOptionValue directly on the solver
+            # allows one to pass advanced options that aren't available
+            # on the HighOptions class (e.g., presolve_rule_off)
+            if solver.setOptionValue(name, value) == hp.HighsStatus.kError:
+                raise ValueError(
+                    f"HIGHS returned status kError for option (name, value): ({name}, {value})"
+                )
 
-        if options.write_model_file:
+        if write_model_file:
             # TODO: Names can be collected upstream more systematically
             # (or in the parent class) to be used by all solvers.
             column_names = []
@@ -275,14 +296,12 @@ class HIGHS(ConicSolver):
                 collect_column_names(variable, column_names)
             lp.col_names_ = column_names
 
-        solver = hp.Highs()
-        solver.passOptions(options)
         solver.passModel(model)
 
-        if options.write_model_file:
+        if write_model_file:
             # TODO: This part can be removed once the following HiGS PR is released:
             # https://github.com/ERGO-Code/HiGHS/pull/2274
-            solver.writeModel(options.write_model_file)
+            solver.writeModel(write_model_file)
 
         if warm_start and solver_cache is not None and self.name() in solver_cache:
             old_solver, old_data, old_result = solver_cache[self.name()]
