@@ -18,13 +18,16 @@ import numpy as np
 import scipy  # For version checks
 
 import cvxpy.settings as s
+
+# Conic interfaces
 from cvxpy.reductions.solvers.conic_solvers.cbc_conif import CBC as CBC_con
 from cvxpy.reductions.solvers.conic_solvers.clarabel_conif import CLARABEL as CLARABEL_con
 from cvxpy.reductions.solvers.conic_solvers.copt_conif import COPT as COPT_con
+from cvxpy.reductions.solvers.conic_solvers.cosmo_conif import COSMO as COSMO_con
 from cvxpy.reductions.solvers.conic_solvers.cplex_conif import CPLEX as CPLEX_con
+from cvxpy.reductions.solvers.conic_solvers.cuclarabel_conif import CUCLARABEL as CUCLARABEL_con
+from cvxpy.reductions.solvers.conic_solvers.cuopt_conif import CUOPT as CUOPT_con
 from cvxpy.reductions.solvers.conic_solvers.cvxopt_conif import CVXOPT as CVXOPT_con
-
-# Conic interfaces
 from cvxpy.reductions.solvers.conic_solvers.diffcp_conif import DIFFCP as DIFFCP_con
 from cvxpy.reductions.solvers.conic_solvers.ecos_bb_conif import ECOS_BB as ECOS_BB_con
 from cvxpy.reductions.solvers.conic_solvers.ecos_conif import ECOS as ECOS_con
@@ -33,6 +36,7 @@ from cvxpy.reductions.solvers.conic_solvers.glpk_conif import GLPK as GLPK_con
 from cvxpy.reductions.solvers.conic_solvers.glpk_mi_conif import GLPK_MI as GLPK_MI_con
 from cvxpy.reductions.solvers.conic_solvers.gurobi_conif import GUROBI as GUROBI_con
 from cvxpy.reductions.solvers.conic_solvers.highs_conif import HIGHS as HIGHS_con
+from cvxpy.reductions.solvers.conic_solvers.knitro_conif import KNITRO as KNITRO_con
 from cvxpy.reductions.solvers.conic_solvers.mosek_conif import MOSEK as MOSEK_con
 from cvxpy.reductions.solvers.conic_solvers.nag_conif import NAG as NAG_con
 from cvxpy.reductions.solvers.conic_solvers.pdlp_conif import PDLP as PDLP_con
@@ -42,26 +46,30 @@ from cvxpy.reductions.solvers.conic_solvers.scipy_conif import SCIPY as SCIPY_co
 from cvxpy.reductions.solvers.conic_solvers.scs_conif import SCS as SCS_con
 from cvxpy.reductions.solvers.conic_solvers.sdpa_conif import SDPA as SDPA_con
 from cvxpy.reductions.solvers.conic_solvers.xpress_conif import XPRESS as XPRESS_con
-from cvxpy.reductions.solvers.qp_solvers.copt_qpif import COPT as COPT_qp
-from cvxpy.reductions.solvers.qp_solvers.cplex_qpif import CPLEX as CPLEX_qp
 
 # QP interfaces
+from cvxpy.reductions.solvers.qp_solvers.copt_qpif import COPT as COPT_qp
+from cvxpy.reductions.solvers.qp_solvers.cplex_qpif import CPLEX as CPLEX_qp
 from cvxpy.reductions.solvers.qp_solvers.daqp_qpif import DAQP as DAQP_qp
 from cvxpy.reductions.solvers.qp_solvers.gurobi_qpif import GUROBI as GUROBI_qp
 from cvxpy.reductions.solvers.qp_solvers.highs_qpif import HIGHS as HIGHS_qp
+from cvxpy.reductions.solvers.qp_solvers.knitro_qpif import KNITRO as KNITRO_qp
 from cvxpy.reductions.solvers.qp_solvers.mpax_qpif import MPAX as MPAX_qp
 from cvxpy.reductions.solvers.qp_solvers.osqp_qpif import OSQP as OSQP_qp
 from cvxpy.reductions.solvers.qp_solvers.piqp_qpif import PIQP as PIQP_qp
 from cvxpy.reductions.solvers.qp_solvers.proxqp_qpif import PROXQP as PROXQP_qp
+from cvxpy.reductions.solvers.qp_solvers.qpalm_qpif import QPALM as QPALM_qp
 from cvxpy.reductions.solvers.qp_solvers.xpress_qpif import XPRESS as XPRESS_qp
 from cvxpy.utilities.versioning import Version
 
 solver_conic_intf = [DIFFCP_con(), ECOS_con(),
                      CVXOPT_con(), GLPK_con(), COPT_con(),
-                     GLPK_MI_con(), CBC_con(), CLARABEL_con(), SCS_con(), SDPA_con(),
+                     GLPK_MI_con(), CBC_con(), CLARABEL_con(), COSMO_con(), SCS_con(), SDPA_con(),
                      GUROBI_con(), MOSEK_con(), CPLEX_con(), NAG_con(), XPRESS_con(),
                      SCIP_con(), SCIPY_con(), HIGHS_con(), GLOP_con(), PDLP_con(),
-                     QOCO_con(), ECOS_BB_con()]
+                     QOCO_con(), CUCLARABEL_con(), CUOPT_con(), ECOS_BB_con(),
+                     KNITRO_con()]
+
 solver_qp_intf = [OSQP_qp(),
                   GUROBI_qp(),
                   CPLEX_qp(),
@@ -69,9 +77,11 @@ solver_qp_intf = [OSQP_qp(),
                   COPT_qp(),
                   PIQP_qp(),
                   PROXQP_qp(),
+                  QPALM_qp(),
                   DAQP_qp(),
                   HIGHS_qp(),
                   MPAX_qp(),
+                  KNITRO_qp(),
                   ]
 
 SOLVER_MAP_CONIC = {solver.name(): solver for solver in solver_conic_intf}
@@ -83,8 +93,9 @@ SOLVER_MAP_QP = {solver.name(): solver for solver in solver_qp_intf}
 CONIC_SOLVERS = [s.MOSEK, s.CLARABEL, s.SCS, s.ECOS, s.SDPA,
                  s.CPLEX, s.GUROBI, s.COPT, s.GLPK, s.NAG,
                  s.GLPK_MI, s.CBC, s.CVXOPT, s.XPRESS, s.DIFFCP,
-                 s.SCIP, s.SCIPY, s.HIGHS, s.GLOP, s.PDLP, s.QOCO, 
-                 s.ECOS_BB]
+                 s.SCIP, s.SCIPY, s.HIGHS, s.GLOP, s.PDLP, s.QOCO,
+                 s.CUCLARABEL, s.CUOPT, s.ECOS_BB, s.KNITRO, s.COSMO]
+
 QP_SOLVERS = [s.OSQP,
               s.GUROBI,
               s.CPLEX,
@@ -93,13 +104,16 @@ QP_SOLVERS = [s.OSQP,
               s.COPT,
               s.PIQP,
               s.PROXQP,
+              s.QPALM,
               s.DAQP,
-              s.MPAX]
+              s.MPAX,
+              s.KNITRO]
 DISREGARD_CLARABEL_SDP_SUPPORT_FOR_DEFAULT_RESOLUTION = True
 MI_SOLVERS = [s.GLPK_MI, s.MOSEK, s.GUROBI, s.CPLEX,
-              s.XPRESS, s.CBC, s.SCIP, s.HIGHS, s.COPT, s.ECOS_BB]
+              s.XPRESS, s.CBC, s.SCIP, s.HIGHS, s.COPT,
+              s.CUOPT, s.ECOS_BB, s.KNITRO]
 MI_SOCP_SOLVERS = [s.MOSEK, s.GUROBI, s.CPLEX, s.XPRESS,
-                   s.SCIP, s.ECOS_BB]
+                   s.SCIP, s.ECOS_BB, s.KNITRO]
 
 # Acknowledge MI solver support for SciPy >= 1.9.
 if not (Version(scipy.__version__) < Version('1.9.0')):
