@@ -74,6 +74,7 @@ GLPK_MI = "GLPK_MI"
 GLOP = "GLOP"
 CBC = "CBC"
 COPT = "COPT"
+CUOPT = "CUOPT"
 ECOS = "ECOS"
 ECOS_BB = "ECOS_BB"
 SCS = "SCS"
@@ -84,6 +85,7 @@ OSQP = "OSQP"
 PIQP = "PIQP"
 PROXQP = "PROXQP"
 QOCO = "QOCO"
+QPALM = "QPALM"
 CPLEX = "CPLEX"
 MOSEK = "MOSEK"
 XPRESS = "XPRESS"
@@ -92,13 +94,17 @@ PDLP = "PDLP"
 SCIP = "SCIP"
 SCIPY = "SCIPY"
 CLARABEL = "CLARABEL"
+CUCLARABEL = "CUCLARABEL"
 DAQP = "DAQP"
 HIGHS = "HIGHS"
 MPAX = "MPAX"
+KNITRO = "KNITRO"
+COSMO = "COSMO"
 SOLVERS = [CLARABEL, ECOS, CVXOPT, GLOP, GLPK, GLPK_MI,
            SCS, SDPA, GUROBI, OSQP, CPLEX,
-           MOSEK, CBC, COPT, XPRESS, PIQP, PROXQP, QOCO,
-           NAG, PDLP, SCIP, SCIPY, DAQP, HIGHS, MPAX]
+           MOSEK, CBC, COPT, XPRESS, PIQP, PROXQP, QOCO, QPALM,
+           NAG, PDLP, SCIP, SCIPY, DAQP, HIGHS, MPAX,
+           CUCLARABEL, CUOPT, KNITRO, COSMO]
 
 # Xpress-specific items
 XPRESS_IIS = "XPRESS_IIS"
